Create README.md
This project displays a python program that finds the most extreme gradients on which to set up snowboard runs for an extreme sports holiday park. The program pulls in hillslope heights from a raster text file containing heights of the area, calculates the maximum slope in all the raster cells using D8 algorithm, builds datasets with slope gradients rather than the heights, displays the heights and gradients as an image and save them in similar formats.